add code editor component
