+++ componenets/index.tsx
@@ -1,0 +1,1 @@
+export { default as CodeEditorComponent } from "./CodeEditor.component";
